Wait for all chunks to arrive
This waits for all chunks to arrive after the call to '/sota/finish'. It
assumes, that all messages will arrive eventually, no message may ever
be lost, otherwise the file descriptor for the package is never closed.

Also adds a trait called 'Unpack' to safely retrieve an entry from a
HashMap, printing an error message to stdout if the entry can't be
found. It is used to hide the HashMapEntry unwrapping and error
handling.