Fix WS browser end not getting events
Two interesting things were at play here -- first, we moved the Mutex from
Gateway closer to where the locked resource is (in Websocket really), and
required the implementer to be Sync. Before that, we were locking the whole
implementer in the reading thread 99% of the time, and so the writer thread
could never acquire the lock.

The other problem was in the Websocket struct -- mapping over a
HashMap with a side-effecting function is apparently optimized away, which was
fixed by turning it into a for.